booster api renamed
