+++ src/index.js
@@ -1,7 +1,7 @@
 import './js/hero-filter.js';
 import './js/hero-search.js';
 import './js/booster.js';
-import './js/choose-country.js';
+import './js/booster-api.js';
 import './js/buy-tickets.js';
 import './js/modal.js';
 import './js/modal-team.js';
